ai_analyzer-main/requirements.txt
streamlit
pandas
plotly
networkx

langchain
langchain-community
langchain-groq
langchain-text-splitters
duckduckgo-search
tavily-python
requests

faiss-cpu

pypdf
pdfplumber
docx2txt
sentence-transformers
python-dotenv
langchain_text_splitters